% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{addToMetadata}
\alias{addToMetadata}
\alias{addToMetadata,SummarySet-method}
\title{Add to metadata in the SummarySet}
\usage{
addToMetadata(
  summary_set,
  id = getMetadata(summary_set)$id,
  sample_size = getMetadata(summary_set)$sample_size,
  nsnp = getMetadata(summary_set)$nsnp,
  trait = getMetadata(summary_set)$trait,
  sd = getMetadata(summary_set)$sd,
  unit = getMetadata(summary_set)$unit,
  ncontrol = getMetadata(summary_set)$ncontrol,
  build = getMetadata(summary_set)$build,
  population = getMetadata(summary_set)$population,
  ncase = getMetadata(summary_set)$ncase
)

\S4method{addToMetadata}{SummarySet}(
  summary_set,
  id = getMetadata(summary_set)$id,
  sample_size = getMetadata(summary_set)$sample_size,
  nsnp = getMetadata(summary_set)$nsnp,
  trait = getMetadata(summary_set)$trait,
  sd = getMetadata(summary_set)$sd,
  unit = getMetadata(summary_set)$unit,
  ncontrol = getMetadata(summary_set)$ncontrol,
  build = getMetadata(summary_set)$build,
  population = getMetadata(summary_set)$population,
  ncase = getMetadata(summary_set)$ncase
)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object.}

\item{id}{GWAS study ID.}

\item{sample_size}{Sample size.}

\item{nsnp}{Number of variants in the study.}

\item{trait}{Phenotype name corresponding the the variant.}

\item{sd}{Trait standard deviation.}

\item{unit}{Unit.}

\item{ncontrol}{Number of controls in study.}

\item{build}{genome build version.}

\item{population}{Study sample population.}

\item{ncase}{Number of cases in study.}
}
\value{
gwasglue2 SummarySet object with metadata stored.
}
\description{
Add to metadata in the SummarySet
}
