% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{assertSameShape}
\alias{assertSameShape}
\alias{assertSameShape,DataSet-method}
\title{Assert if the shapes of SummarySets in the gwasglue2 DataSet are the same}
\usage{
assertSameShape(dataset)

\S4method{assertSameShape}{DataSet}(dataset)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object.}
}
\value{
logical TRUE/FALSE
}
\description{
Assert if the shapes of SummarySets in the gwasglue2 DataSet are the same
}
