% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{clumpTophits}
\alias{clumpTophits}
\title{A function to clump the top hit variants}
\usage{
clumpTophits(traits, clump = TRUE, source = ieugwasr::check_access_token())
}
\arguments{
\item{traits}{ID of GWAS studies to query}

\item{clump}{Logical (default TRUE)}

\item{source}{IEU (default)}
}
\value{
Array of SNPs rsids
}
\description{
A function to clump the top hit variants
}
