% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTo.R
\docType{methods}
\name{convertForTwoSampleMR}
\alias{convertForTwoSampleMR}
\alias{convertForTwoSampleMR,DataSet-method}
\title{Convert tool to TwoSampleMR format}
\usage{
convertForTwoSampleMR(dataset)

\S4method{convertForTwoSampleMR}{DataSet}(dataset)
}
\arguments{
\item{dataset}{The gwasglue2 DataSet object}
}
\value{
The gwasglue2 SummarySet object converted to TwoSampleMR format
}
\description{
Converts SummarySets within a Dataset to a format that can be read by TwoSampleMR
}
