% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTo.R
\docType{methods}
\name{convertToTwoSampleMR}
\alias{convertToTwoSampleMR}
\alias{convertToTwoSampleMR,DataSet-method}
\title{Convert tool to TwoSampleMR format}
\usage{
convertToTwoSampleMR(dataset)

\S4method{convertToTwoSampleMR}{DataSet}(dataset)
}
\arguments{
\item{dataset}{The gwasglue2 DataSet object}
}
\value{
A TwoSampleMR data.frame input format
}
\description{
Converts SummarySets within a Dataset to a format that can be read by TwoSampleMR
}
