% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_dataset}
\alias{create_dataset}
\title{Creates a DataSet object using gwasglue2 SummarySet objects, and harmonise data against data}
\usage{
create_dataset(
  data = list(),
  harmonise = TRUE,
  tolerance = 0.08,
  action = 1,
  strand = "forward"
)
}
\arguments{
\item{data}{A list of gwasglue2 SummarySet objects}

\item{harmonise}{logical (default TRUE). It harmonises the summary sets in the DataSet against each other.}

\item{tolerance}{Inherited from harmoniseData() (default 0.08)}

\item{action}{Inherited from harmoniseData() (Default 1)
\itemize{
\item \code{action = 1}: Assume all alleles are coded on the forward strand, i.e. do not attempt to flip alleles
\item \code{action = 2}: Try to infer positive strand alleles, using allele frequencies for palindromes (default, conservative);
\item \code{action = 3}: Correct strand for non-palindromic SNPs, and drop all palindromic SNPs from the analysis (more conservative).
}}

\item{strand}{dna strand orientation  (Default "forward", other option "reverse")}
}
\value{
A harmonised gwasglue2 DataSet object
}
\description{
Creates a DataSet object using gwasglue2 SummarySet objects, and harmonise data against data
}
