% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{create_meta_summaryset}
\alias{create_meta_summaryset}
\title{Meta analysis}
\usage{
create_meta_summaryset(dataset)
}
\arguments{
\item{dataset}{gwasglue2 DataSet object}
}
\value{
gwasglue2 SummarySet object
}
\description{
Statistical combination of the
results from two or more separate studies.
It uses the fixed-effect model assuming that one true
effect size underlies all the studies in the meta-analysis.
}
