% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_metadata}
\alias{create_metadata}
\title{Metadata object}
\usage{
create_metadata(
  metadata = NULL,
  id = NA,
  sample_size = NA,
  nsnp = NA,
  trait = NA,
  sd = NA,
  unit = NA,
  ncontrol = NA,
  build = NA,
  population = NA,
  ncase = NA,
  ...
)
}
\arguments{
\item{metadata}{A dataframe with metadata information. Not required.}

\item{id}{GWAS study ID.}

\item{sample_size}{Sample size.}

\item{nsnp}{Number of variants in the study.}

\item{trait}{Phenotype name corresponding the the variant.}

\item{sd}{Trait standard deviation.}

\item{unit}{Unit}

\item{ncontrol}{Nb of controls in study}

\item{build}{genome build version.}

\item{population}{Study sample population.}

\item{ncase}{Number of cases in study.}

\item{...}{Other metadata information}
}
\value{
A metadata list.
}
\description{
Reads metadata and converts it to gwasglue2 format.
}
