% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{create_summary_set_from_lbf}
\alias{create_summary_set_from_lbf}
\title{Create SummarySet from log Bayes Factor}
\usage{
create_summary_set_from_lbf(summaryset, lbf, L)
}
\arguments{
\item{summaryset}{gwasglue2 SummarySet object}

\item{lbf}{p-vector of log Bayes Factors for each SNP}

\item{L}{credible set index number}
}
\value{
marginalised summaryset (beta, se and trait id)
}
\description{
Create SummarySet from log Bayes Factor
}
