% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_summaryset}
\alias{create_summaryset}
\title{A function to create a gwasglue2 SummarySet object from different sources and formats}
\usage{
create_summaryset(
  data,
  metadata = NULL,
  type = "tibble",
  qc = FALSE,
  beta_col = NULL,
  se_col = NULL,
  samplesize_col = NULL,
  pvalue_col = NULL,
  logpvalue_col = NULL,
  chr_col = NULL,
  position_col = NULL,
  rsid_col = NULL,
  effect_allele_col = NULL,
  other_allele_col = NULL,
  eaf_col = NULL,
  id_col = NULL,
  trait_col = NULL
)
}
\arguments{
\item{data}{GWAS summary statistics.}

\item{metadata}{A list with metadata information. If NULL, it creates metadata with information retrieved from the dataset}

\item{type}{Input @param data type. Default is \code{"tibble"}. Other options: \code{"vcf"}}

\item{qc}{Quality control. It checks the @param data and look for problems that can stop gwasglue2 from runing. If TRUE gwasglue will try to solve the problems.  Default is FALSE}

\item{beta_col}{Name of column with effect sizes. The default is \code{"beta"} for @param type \code{"tibble"} and \code{"ES"}for @param type \code{"vcf"}..}

\item{se_col}{Name of column with standard errors. The default is \code{"se"} for @param type \code{"tibble"} and \code{"SE"}for @param type \code{"vcf"}.}

\item{samplesize_col}{Column name for sample size. The default is \code{"n"} for @param type \code{"tibble"} and \code{"SS"}for @param type \code{"vcf"}.}

\item{pvalue_col}{Name of column with p-value. The default is \code{"p"}.}

\item{logpvalue_col}{Name of column with log(p-value). The default is \code{"LP"} for @param type \code{"vcf"}.}

\item{chr_col}{Column name for chromosome . The default is \code{"chr"} for @param type \code{"tibble"} and \code{"seqnames"}for @param type \code{"vcf"}.}

\item{position_col}{Column name for the position. Together, with @param chr gives the physical coordinates of the variant. The default is \code{"position"} for @param type \code{"tibble"} and \code{"start"}for @param type \code{"vcf"}.}

\item{rsid_col}{Required name of column with variants rs IDs. The default is \code{"rsid"} for @param type \code{"tibble"} and \code{"ID"}for @param type \code{"vcf"}.}

\item{effect_allele_col}{Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"ea"} for @param type \code{"tibble"} and \code{"ALT"}for @param type \code{"vcf"}.}

\item{other_allele_col}{Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \verb{"nea} for @param type \code{"tibble"} and \code{"REF"}for @param type \code{"vcf"}.}

\item{eaf_col}{Name of column with effect allele frequency. The default is \code{"eaf"} for @param type \code{"tibble"} and \code{"AF"}for @param type \code{"vcf"}.}

\item{id_col}{GWAS study ID column. The default is \code{"id"}.}

\item{trait_col}{Column name for the column with phenotype name corresponding the the variant. The default is \code{"trait"}}
}
\value{
A gwasglue2 SummarySet object
}
\description{
A function to create a gwasglue2 SummarySet object from different sources and formats
}
\seealso{
\code{\link[=create_metadata]{create_metadata()}} to create a metadata object
}
