% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_summaryset_from_gwasvcf}
\alias{create_summaryset_from_gwasvcf}
\title{A function to create a gwasglue2 SummarySet object from a vcf file}
\usage{
create_summaryset_from_gwasvcf(
  data,
  metadata = NULL,
  qc = FALSE,
  beta_col = "ES",
  se_col = "SE",
  samplesize_col = "SS",
  logpvalue_col = "LP",
  pvalue_col = "p",
  chr_col = "seqnames",
  position_col = "start",
  rsid_col = "ID",
  effect_allele_col = "ALT",
  other_allele_col = "REF",
  eaf_col = "AF",
  id_col = "id",
  build = NULL
)
}
\arguments{
\item{data}{GWAS summary statistics. In the GWAS vcf dataframe format}

\item{metadata}{A list with metadata information. If NULL, it creates metadata with information retrieved from the dataset}

\item{qc}{Quality control. It checks the @param data and look for problems that can stop gwasglue2 from runing. If TRUE gwasglue will try to solve the problems.  Default is FALSE}

\item{beta_col}{Name of column with effect sizes. The default is \code{"ES"}.}

\item{se_col}{Name of column with standard errors. The default is \code{"SE"}.}

\item{samplesize_col}{Column name for sample size. The default is \code{"SS"}.}

\item{logpvalue_col}{Name of column with log(p-value). The default is \code{"LP"}.}

\item{pvalue_col}{Name of column with p-value. The default is \code{"p"}.}

\item{chr_col}{Column name for chromosome . The default is \code{"seqnames"}.}

\item{position_col}{Column name for the position. Together, with @param chr gives the physical coordinates of the variant. The default is \code{"start"}.}

\item{rsid_col}{Required name of column with variants rs IDs. The default is \code{"ID"}.}

\item{effect_allele_col}{Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"ALT"}.}

\item{other_allele_col}{Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"REF"}.}

\item{eaf_col}{Name of column with effect allele frequency. The default is \code{"AF"}.}

\item{id_col}{GWAS study ID column. The default is \code{"id"}.}

\item{build}{Reference genome assembly to generate the vcf file. Default is NULL.
\itemize{
\item Options are \code{"NCBI34"}, \code{"NCBI35"}, \code{"NCBI36"}, \code{"GRCh37"} or "GRCh38".
}}
}
\value{
A gwasglue2 SummarySet object
}
\description{
A function to create a gwasglue2 SummarySet object from a vcf file
}
\seealso{
\code{\link[=create_metadata]{create_metadata()}} to create a metadata object
}
