% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createDataSet}
\alias{createDataSet}
\title{A function to create a DataSet object and harmonise data against data}
\usage{
createDataSet(
  traits,
  variants,
  tools,
  source = "IEUopenGWAS",
  harmonise = TRUE,
  tolerance = 0.08,
  action = 1,
  strand = "forward"
)
}
\arguments{
\item{traits}{ID of GWAS studies to query}

\item{variants}{Array of variants rsids}

\item{tools}{Array of methods that gwasglue2 is going to convert the summarySet to (eg. "mr")}

\item{source}{ID of Data Source to set the DataSet @slot metadata (Default and only option: IEUopenGWAS).}

\item{harmonise}{logical (default TRUE). It harmonises the summary sets in the DataSet against each other.}

\item{tolerance}{Inherited from harmoniseData() (default 0.08)}

\item{action}{Inherited from harmoniseData() (Default 1)}

\item{strand}{dna strand orientation  (Default "forward", other option "reverse")}
}
\value{
A harmonised gwasglue2 DataSet object
}
\description{
A function to create a DataSet object and harmonise data against data
}
