% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieugwas_utils.R
\name{createSumset}
\alias{createSumset}
\title{A function to call the summary statistics associated with traits and variants chosen}
\usage{
createSumset(traits, variants)
}
\arguments{
\item{variants}{}
}
\description{
A function to call the summary statistics associated with traits and variants chosen
}
