% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\title{An S4 class to represent the Data Set}
\description{
An S4 class to represent the Data Set
}
\section{Slots}{

\describe{
\item{\code{summary_sets}}{A list of SummarySet objects (default NA).}

\item{\code{overlap_SNPs}}{among all SummarySets}

\item{\code{is_resized}}{logical (default FALSE).}

\item{\code{is_harmonised}}{logical (default FALSE).}

\item{\code{ld_matrices}}{A list of LD correlation matrices (default NA).}

\item{\code{is_harmonisedLD}}{logical (default FALSE).}

\item{\code{is_converted}}{logical (default FALSE).}
}}

