% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\alias{DataSet}
\title{An S4 class to represent the Data Set}
\usage{
DataSet(...)
}
\arguments{
\item{...}{Array of gwasglue2 SummarySet object names.}
}
\value{
A gwasglue2 DataSet object
}
\description{
An S4 class to represent the Data Set

DataSet function
}
\section{Slots}{

\describe{
\item{\code{summary_sets}}{A list of SummarySet objects (default NA).}

\item{\code{overlap_variants}}{among all SummarySets}

\item{\code{is_resized}}{logical (default FALSE).}

\item{\code{is_harmonised}}{logical (default FALSE).}

\item{\code{overall_dropped_SNPs}}{A vector of RSIDs that were removed from the summary_sets.}

\item{\code{dropped_SNPs}}{A list of pairwise harmonising output (SNPs removed from the summary_sets )}

\item{\code{palindromic_SNPs}}{A list of pairwise harmonising output.}

\item{\code{ambiguous_SNPs}}{A list of pairwise harmonising output.}

\item{\code{incompatible_alleles_SNPs}}{A list of pairwise harmonising output.}

\item{\code{ld_matrix}}{LD matrix from reference population}

\item{\code{is_harmonisedLD}}{logical (default FALSE).}

\item{\code{zscores}}{vector of calculated z-scores}

\item{\code{susie_marginalised}}{logical (default FALSE).}

\item{\code{susieR}}{susieR::susie_rss() output}

\item{\code{is_converted}}{logical (default FALSE).}

\item{\code{describe}}{A description of the DataSet (default NA).}

\item{\code{trait_organisation}}{A list with the trait organization within the DataSet (default NA).}
}}

