% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dataset_to_hyprcoloc}
\alias{dataset_to_hyprcoloc}
\title{DataSet to hyprcoloc}
\usage{
dataset_to_hyprcoloc(dataset)
}
\arguments{
\item{dataset}{gwasglue2 DataSet object}
}
\value{
parameters needed to run hyprcoloc
}
\description{
dataset_to_hyprcoloc is a wrapper function used inside ritarasteiro/hyprcoloc::hyprcoloc() to read DataSet objects
}
