% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{dimData}
\alias{dimData}
\alias{dimData,SummarySet-method}
\title{Dimensions of the GWAS Summary Statistics data}
\usage{
dimData(object)

\S4method{dimData}{SummarySet}(object)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object}
}
\value{
The dimensions of the GWAS Summary Statistics data
}
\description{
Dimensions of the GWAS Summary Statistics data
}
