% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{download_chainfile}
\alias{download_chainfile}
\title{Chain files downloader}
\usage{
download_chainfile(from = "GRCh37", to = "GRCh38")
}
\arguments{
\item{from}{genome assembly to which GWAS summary data is currently mapped. Default "GRCh37".
\itemize{
\item Other options: "NCBI34", "NCBI35", "NCBI36" and "GRCh38"
}}

\item{to}{genome  assembly to which should be mapped. Default "GRCh38"
\itemize{
\item Other options: "NCBI34", "NCBI35", "NCBI36" and "GRCh37"
}}
}
\value{
A chain file
}
\description{
Human genome chain files are download from UCSC
}
