% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getAttributes}
\alias{getAttributes}
\alias{getAttributes,SummarySet-method}
\title{Get Method to retrieve the attributes linked to the SummarySet}
\usage{
getAttributes(summary_set)

\S4method{getAttributes}{SummarySet}(summary_set)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object}
}
\value{
The attributes associated with the SummarySet
}
\description{
Get Method to retrieve the attributes linked to the SummarySet
}
