% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\docType{methods}
\name{getDataSet}
\alias{getDataSet}
\title{Get Method to retrieve the gwasglue2 DataSet object}
\usage{
getDataSet(dataset, index)
}
\arguments{
\item{index}{Index of gwasglue2 SummarySet objects within DataSet}

\item{list_datasets}{A gwasglue2 ListDataSet object}
}
\value{
A gwasglue2 DataSet object
}
\description{
Get Method to retrieve the gwasglue2 DataSet object
}
