% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmoniseLD_method.R
\docType{methods}
\name{getLDMatrix}
\alias{getLDMatrix}
\alias{getLDMatrix,DataSet-method}
\title{Get Method to retrieve the Linkage Disequilibrium matrix}
\usage{
getLDMatrix(dataset)

\S4method{getLDMatrix}{DataSet}(dataset)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object}
}
\value{
The LD matrix
}
\description{
Get Method to retrieve the Linkage Disequilibrium matrix
}
