% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{getLength}
\alias{getLength}
\alias{getLength,DataSet-method}
\title{Size of the DataSet}
\usage{
getLength(object)

\S4method{getLength}{DataSet}(object)
}
\arguments{
\item{object}{A gwasglue2 DataSet objec}
}
\value{
Number of gwasglue2 SummarySet objects within the DataSet
}
\description{
Size of the DataSet
}
