% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getMRlabel}
\alias{getMRlabel}
\alias{getMRlabel,SummarySet-method}
\title{Get Method to retrieve the Mendelian randomization (MR) label linked to the SummarySet}
\usage{
getMRlabel(object)

\S4method{getMRlabel}{SummarySet}(object)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object}
}
\value{
The MR label associated with the SummarySet
}
\description{
Get Method to retrieve the Mendelian randomization (MR) label linked to the SummarySet
}
