% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getRSID}
\alias{getRSID}
\alias{getRSID,SummarySet-method}
\title{Get Method to retrieve RSID/variants stored in the SummarySet}
\usage{
getRSID(summary_set)

\S4method{getRSID}{SummarySet}(summary_set)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object}
}
\value{
The RSID/variants
}
\description{
Get Method to retrieve RSID/variants stored in the SummarySet
}
\seealso{
Similar to \code{\link[=getVariants]{getVariants()}}
}
