% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getShape}
\alias{getShape}
\alias{getShape,SummarySet-method}
\title{Get the Shape of the gwasglue2 objects}
\usage{
getShape(object)

\S4method{getShape}{SummarySet}(object)
}
\arguments{
\item{object}{A gwasglue2 SummarySet or DataSet object.}
}
\value{
The shape of the gwasglue2 object
}
\description{
Get the Shape of the gwasglue2 objects
}
