% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getSummaryData}
\alias{getSummaryData}
\alias{getSummaryData,SummarySet-method}
\title{Get Method to retrieve the GWAS Summary Statistics  from the SummarySet}
\usage{
getSummaryData(object)

\S4method{getSummaryData}{SummarySet}(object)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object.}
}
\value{
A tibble with GWAS summary statistics
}
\description{
Get Method to retrieve the GWAS Summary Statistics  from the SummarySet
}
