% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getSummarySet,DataSet-method}
\alias{getSummarySet,DataSet-method}
\title{Get Methods for summarySet}
\usage{
\S4method{getSummarySet}{DataSet}(object, index)
}
\arguments{
\item{object}{A DataSet S4 object}

\item{index}{index of SummarySet within DataSet}
}
\value{
summarySet object
}
\description{
Get Methods for summarySet
}
