% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{getSummarySet}
\alias{getSummarySet}
\alias{getSummarySet,DataSet-method}
\title{Get Methods for SummarySet}
\usage{
getSummarySet(object, index)

\S4method{getSummarySet}{DataSet}(object, index)
}
\arguments{
\item{object}{A DataSet S4 object}

\item{index}{index of SummarySet within DataSet}
}
\value{
summarySet object
}
\description{
Get Methods for SummarySet
}
