% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getTool}
\alias{getTool}
\alias{getTool,SummarySet-method}
\title{Get Method to retrieve the tools stored in the SummarySet}
\usage{
getTool(object)

\S4method{getTool}{SummarySet}(object)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object}
}
\value{
The tools
}
\description{
Get Method to retrieve the tools stored in the SummarySet
}
