% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{getTraitOrg}
\alias{getTraitOrg}
\alias{getTraitOrg,DataSet-method}
\title{Get the trait organisation of the gwasglue2 DataSet}
\usage{
getTraitOrg(dataset)

\S4method{getTraitOrg}{DataSet}(dataset)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object.}
}
\value{
The trait organisation of the gwasglue2 object
}
\description{
Get the trait organisation of the gwasglue2 DataSet
}
