% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasglue.R
\name{gwasglue}
\alias{gwasglue}
\title{gwasglue  (WIP)}
\usage{
gwasglue(yaml, read = FALSE, path_to_yaml = "config.yaml")
}
\arguments{
\item{yaml}{A yaml R object.}

\item{read}{A boolean indicating whether the yaml object is a path to a yaml file or a yaml R object. IF TRUE, yaml is a path.}

\item{path_to_yaml}{A character vector indicating the path to the yaml file.}
}
\value{
A gwasglue2 DataSet object.
}
\description{
gwasglue  (WIP)
}
