% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise}
\alias{harmonise}
\title{Harmonise the alleles and effects between two summary sets
This is based o the harmonise function of TwoSampleMR}
\usage{
harmonise(rsid, A1, A2, B1, B2, betaA, betaB, fA, fB, tolerance, action)
}
\arguments{
\item{action}{Level of strictness in dealing with SNPs.
\itemize{
\item \code{action = 1}: Assume all alleles are coded on the forward strand, i.e. do not attempt to flip alleles
\item \code{action = 2}: Try to infer positive strand alleles, using allele frequencies for palindromes (default, conservative);
\item \code{action = 3}: Correct strand for non-palindromic SNPs, and drop all palindromic SNPs from the analysis (more conservative).
}}
}
\description{
Harmonise the alleles and effects between two summary sets
This is based o the harmonise function of TwoSampleMR
}
