% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_method.R
\docType{methods}
\name{harmoniseData}
\alias{harmoniseData}
\alias{harmoniseData,DataSet-method}
\title{Harmonise the alleles and effects between two summary sets}
\usage{
harmoniseData(dataset, tolerance, action)

\S4method{harmoniseData}{DataSet}(dataset, tolerance = 0.08, action = 2)
}
\arguments{
\item{dataset}{The gwasglue2 DataSet object}

\item{tolerance}{Tolerance value.}

\item{action}{Level of strictness in dealing with SNPs.
\itemize{
\item \code{action = 1}: Assume all alleles are coded on the forward strand, i.e. do not attempt to flip alleles
\item \code{action = 2}: Try to infer positive strand alleles, using allele frequencies for palindromes (default, conservative);
\item \code{action = 3}: Correct strand for non-palindromic SNPs, and drop all palindromic SNPs from the analysis (more conservative).
}}
}
\value{
The gwasglue2 DataSet object harmonised
}
\description{
Harmonise the alleles and effects between two summary sets
}
