% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_method.R
\docType{methods}
\name{isHarmonised}
\alias{isHarmonised}
\alias{isHarmonised,DataSet-method}
\title{Check if the DataSet is harmonised}
\usage{
isHarmonised(dataset)

\S4method{isHarmonised}{DataSet}(dataset)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object}
}
\value{
TRUE/FALSE
}
\description{
Check if the DataSet is harmonised
}
