% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmoniseLD_method.R
\docType{methods}
\name{isHarmonisedLD}
\alias{isHarmonisedLD}
\alias{isHarmonisedLD,DataSet-method}
\title{Check if the DataSet is harmonised against LD matrix}
\usage{
isHarmonisedLD(dataset)

\S4method{isHarmonisedLD}{DataSet}(dataset)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object}
}
\value{
TRUE/FALSE
}
\description{
Check if the DataSet is harmonised against LD matrix
}
