% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_constructors.R
\name{job_coloc}
\alias{job_coloc}
\title{YAML constructor for the coloc job}
\usage{
job_coloc(
  traits,
  region,
  build = "GRCh37",
  bfile = NULL,
  plink_bin = NULL,
  write = FALSE,
  outfile = "config.yaml"
)
}
\arguments{
\item{traits}{A character vector of traits to be analysed.}

\item{region}{A character vector of regions to be analysed.}

\item{build}{A character vector  of the reference genome assemblies to generate the vcf file. Default is "GRCh37".
\itemize{
\item Options are \code{"NCBI34"}, \code{"NCBI35"}, \code{"NCBI36"}, \code{"GRCh37"} or "GRCh38".
}}

\item{bfile}{It corresponds to the path and prefix of the plink files used to build the LD correlation matrix.}

\item{plink_bin}{Path to the plink executable}

\item{write}{A boolean indicating whether the yaml object should be written to a file.}

\item{outfile}{A character vector indicating the path to write the yaml file.}
}
\value{
A yaml object.
}
\description{
YAML constructor for the coloc job
}
