% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_constructors.R
\name{job_mr}
\alias{job_mr}
\title{YAML constructor for the MR job}
\usage{
job_mr(
  exposure,
  outcome,
  tophits = TRUE,
  build = "GRCh37",
  write = FALSE,
  outfile = "config.yaml"
)
}
\arguments{
\item{exposure}{A character vector of exposure traits to be analysed.}

\item{outcome}{A character vector of outcome traits to be analysed.}

\item{tophits}{Logical. Default TRUE. If TRUE, the top hits from the exposure traits are used as the variants to be analysed.}

\item{write}{A boolean indicating whether the yaml object should be written to a file.}

\item{outfile}{A character vector indicating the path to write the yaml file.}
}
\value{
A yaml object.
}
\description{
YAML constructor for the MR job
}
