% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{lbf_to_z_cont}
\alias{lbf_to_z_cont}
\title{Convert log Bayes Factor to summary stats}
\usage{
lbf_to_z_cont(lbf, n, af, prior_v = 50)
}
\arguments{
\item{lbf}{p-vector of log Bayes Factors for each SNP}

\item{n}{Overall sample size}

\item{af}{p-vector of allele frequencies for each SNP}

\item{prior_v}{Variance of prior distribution. SuSiE uses 50}
}
\value{
tibble with lbf, af, beta, se, z
}
\description{
Convert log Bayes Factor to summary stats
}
