% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_ld.R
\name{ld_matrix_local}
\alias{ld_matrix_local}
\title{modified from ieugwasr::ld_matrix_local ()
Get LD matrix using local plink binary and reference dataset}
\usage{
ld_matrix_local(variants, bfile, plink_bin)
}
\arguments{
\item{variants}{List of variants (in plink 'set range' format).}

\item{bfile}{Path to bed/bim/fam ld reference panel}

\item{plink_bin}{Specify path to plink binary. Default = \code{NULL}.}
}
\value{
data frame
}
\description{
modified from ieugwasr::ld_matrix_local ()
Get LD matrix using local plink binary and reference dataset
}
