% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\docType{class}
\name{ListDataSets-class}
\alias{ListDataSets-class}
\alias{ListDataSets}
\title{An S4 class to represent the ListDataSet object}
\usage{
ListDataSets(...)
}
\arguments{
\item{...}{Array of gwasglue2 DataSett object names.}
}
\value{
A gwasglue2 ListDataSet object
}
\description{
An S4 class to represent the ListDataSet object

ListDataSets function
}
\section{Slots}{

\describe{
\item{\code{datasets}}{A list of DataSet objects (default NA).}
}}

