% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Meta analysis}
\usage{
meta_analysis(dataset, method = "fixed")
}
\arguments{
\item{dataset}{gwasglue2 DataSet object}

\item{method}{Uses fixed-effect model. Default ('"fixed"')}
}
\value{
gwasglue2 SummarySet object
}
\description{
Statistical combination of the
results from two or more separate studies.
It uses the fixed-effect model assuming that one true
effect size underlies all the studies in the meta-analysis.
}
