% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_ld.R
\name{mv_harmonise_data}
\alias{mv_harmonise_data}
\title{Harmonise dat1 and dat2 for multivariable MR}
\usage{
mv_harmonise_data(dat1, dat2, harmonise_strictness = 2)
}
\arguments{
\item{dat1}{Output from \code{\link{mv_extract_dat1s}}.}

\item{dat2}{Output from \code{extract_dat2a(dat1$rsid, id_output)}.}

\item{harmonise_strictness}{See the \code{action} option of \code{\link{harmonise_data}}. The default is \code{2}.}
}
\value{
List of vectors and matrices required for mv analysis.
\describe{
\item{dat1_beta}{a matrix of beta coefficients, in which rows correspond to SNPs and columns correspond to dat1s.}
\item{dat1_se}{is the same as \code{dat1_beta}, but for standard errors.}
\item{dat1_pval}{the same as \code{dat1_beta}, but for p-values.}
\item{expname}{A data frame with two variables, \code{id.dat1} and \code{dat1} which are character strings.}
\item{dat2_beta}{an array of effects for the dat2, corresponding to the SNPs in dat1_beta.}
\item{dat2_se}{an array of standard errors for the dat2.}
\item{dat2_pval}{an array of p-values for the dat2.}
\item{outname}{A data frame with two variables, \code{id.dat2} and \code{dat2} which are character strings.}
}
}
\description{
Harmonise dat1 and dat2 for multivariable MR
}
