% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_method.R
\docType{methods}
\name{overlapVariants}
\alias{overlapVariants}
\alias{overlapVariants,DataSet-method}
\title{Look for overlapped variants between SummarySets in the DataSet and Resize}
\usage{
overlapVariants(object, strand)

\S4method{overlapVariants}{DataSet}(object, strand = "forward")
}
\arguments{
\item{object}{The gwasglue2 DataSet object}

\item{strand}{The direction of the DNA sequence \code{forward} or \code{reverse} (Default \code{forward})}
}
\value{
The gwasglue2 DataSet object resized
}
\description{
Look for overlapped variants between SummarySets in the DataSet and Resize
}
