% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{phewasIDs}
\alias{phewasIDs}
\title{PHEWAS_ids}
\usage{
phewasIDs(variants, batch, pval)
}
\arguments{
\item{variants}{Array of variants}

\item{batch}{Vector of batch IDs to search across. If \code{c()} (default) then returns all batches. Iherited fromieugwasr::phewas  # nolint: line_length_linter.}

\item{pval}{p-value threshold. Default = \code{0.00001}. Inherited from ieugwasr::phewas # nolint: line_length_linter.}
}
\value{
id_list Array with traits ids
}
\description{
PHEWAS_ids
}
