% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setAttributes}
\alias{setAttributes}
\alias{setAttributes,SummarySet-method}
\title{Set Method to store the attributes of the SummarySet}
\usage{
setAttributes(summary_set, mr_label = NULL, ...)

\S4method{setAttributes}{SummarySet}(summary_set, mr_label = NULL, ...)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object}

\item{mr_label}{It can be either \code{"exposure"} or `"outcome". Default NULL.}

\item{...}{Other attributes information}
}
\value{
The gwasglue2 SummarySet object with the attributes stored
}
\description{
Set Method to store the attributes of the SummarySet
}
