% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{setChisq}
\alias{setChisq}
\alias{setChisq,DataSet-method}
\title{Calculating Chi squares}
\usage{
setChisq(dataset)

\S4method{setChisq}{DataSet}(dataset)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object}
}
\value{
An extra '"CHISQ"' column in the  GWAS summary statistics tibble.
}
\description{
Calculating Chi squares
}
