% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setMRlabel}
\alias{setMRlabel}
\alias{setMRlabel,SummarySet-method}
\title{Set Method to store the Mendelian randomization (MR) label in the SummarySet}
\usage{
setMRlabel(object, mr_label)

\S4method{setMRlabel}{SummarySet}(object, mr_label)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object}

\item{mr_label}{MR label for SummarySet. It can be either \code{"exposure"} or \code{"outcome"}}
}
\value{
The gwasglue2 SummarySet object with the MR labels stored
}
\description{
Set Method to store the Mendelian randomization (MR) label in the SummarySet
}
