% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setShape}
\alias{setShape}
\alias{setShape,SummarySet-method}
\title{Set the Shape of the gwasglue2 objects}
\usage{
setShape(object, shape)

\S4method{setShape}{SummarySet}(object, shape)
}
\arguments{
\item{object}{A gwasglue2 SummarySet or DataSet object}

\item{shape}{The shape of the GWAS data}
}
\value{
The gwasglue2 object with the shape stored
}
\description{
Set the Shape of the gwasglue2 objects
}
