% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setTool}
\alias{setTool}
\alias{setTool,SummarySet-method}
\title{Set Method to store the tools that gwasglue2 is going to convert the SummarySet to}
\usage{
setTool(object, tools)

\S4method{setTool}{SummarySet}(object, tools)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object}

\item{tools}{The tools}
}
\value{
The gwasglue2 SummarySet object with the tools stored
}
\description{
Set Method to store the tools that gwasglue2 is going to convert the SummarySet to
}
