% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{setTraitOrg}
\alias{setTraitOrg}
\alias{setTraitOrg,DataSet-method}
\title{Set the trait organisation of the gwasglue2 DataSet}
\usage{
setTraitOrg(dataset, ...)

\S4method{setTraitOrg}{DataSet}(dataset, ...)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object}

\item{...}{The organisation of the DataSet}
}
\value{
The gwasglue2 object with the trait organisation stored
}
\description{
Set the trait organisation of the gwasglue2 DataSet
}
