% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setVariantid}
\alias{setVariantid}
\alias{setVariantid,SummarySet-method}
\title{Set Method to create an internal Variant ID for the SummarySet}
\usage{
setVariantid(object)

\S4method{setVariantid}{SummarySet}(object)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object}
}
\value{
A extra '"variantid"' column in the  GWAS summary statistics tibble. The @rdname getSummaryData-methods can be used to retrieve it.
}
\description{
Set Method to create an internal Variant ID for the SummarySet
}
