% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setVariants}
\alias{setVariants}
\alias{setVariants,SummarySet-method}
\title{Set Method to store RSID/variants in the SummarySet}
\usage{
setVariants(object, variants)

\S4method{setVariants}{SummarySet}(object, variants)
}
\arguments{
\item{object}{A gwasglue2 SummarySet object}

\item{variants}{The RSID/variants associated with the GWAS summary statistics}
}
\value{
The gwasglue2 SummarySet object with RSID/variants stored
}
\description{
Set Method to store RSID/variants in the SummarySet
}
\seealso{
Similar to \code{\link[=setRSID]{setRSID()}}
}
