% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{setZscores}
\alias{setZscores}
\alias{setZscores,DataSet-method}
\title{Calculating Z-scores}
\usage{
setZscores(dataset)

\S4method{setZscores}{DataSet}(dataset)
}
\arguments{
\item{dataset}{A gwasglue2 DataSet object}
}
\value{
An extra '"Z"' column in the  GWAS summary statistics tibble.
}
\description{
Calculating Z-scores
}
