% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{class}
\name{SummarySet-class}
\alias{SummarySet-class}
\title{An S4 class to represent the Summary Set}
\description{
An S4 class to represent the Summary Set
}
\section{Slots}{

\describe{
\item{\code{ss}}{A tibble with the GWAS summary statistics (default NA).}

\item{\code{metadata}}{A list with the metadata associated to ss (default NA).}

\item{\code{variants}}{The RSID/variants associated with ss (default NA).}

\item{\code{attributes}}{Attributes of the SummarySet. Eg. MR label Exposure/Outcome (default NA).}

\item{\code{shape}}{The shape of the SummarySet (default NA).}

\item{\code{shape}}{The shape of the SummarySet (default NA).
\itemize{
\item \code{"single"}: single region
\item \code{"multiple"}: multiple regions
\item \code{"independent"}: independent/scattered variants
\item \code{"pruned"}: genome wide - pruned
\item \code{"full"}: genome wide - full
}}
}}

