% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\name{SummarySet}
\alias{SummarySet}
\title{SummarySet function}
\usage{
SummarySet(ss, traits, variants, tools)
}
\arguments{
\item{ss}{It uses createSumset function to call ieugwasR and fill @slot ss}

\item{tools}{}
}
\description{
SummarySet function
}
