% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{susie_to_dataset}
\alias{susie_to_dataset}
\title{SusieR to DataSet}
\usage{
susie_to_dataset(summaryset, s, R)
}
\arguments{
\item{summaryset}{gwasglue2 SummarySet object}

\item{s}{susieR object}

\item{R}{lD matrix}
}
\value{
DataSet object
}
\description{
#' susie_to_dataset is a wrapper function used inside ritarasteiro/susieR::susie_rss() to create a marginalised DataSet object
Converts ABFs to summary statistics and creates a new SummarySet for each credible set. Returns object is a gwasglue2 DataSet class object.
}
