% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{write_ldsc}
\alias{write_ldsc}
\title{Write Summary data into GenomicSEM input files to disk}
\usage{
write_ldsc(dataset, path_to_save = "ldsc")
}
\arguments{
\item{dataset}{gwasglue2 DataSet object}

\item{path_to_save}{character path to save the files. Default is "ldsc"}
}
\value{
character vector of paths to the summary data files. It is the same length as the number of traits in the dataset. It also saves the  compressed 'trait_id.gz' files to the directory specified in path_to_save.
}
\description{
Write Summary data into GenomicSEM input files to disk
}
