% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_summaryset}
\alias{create_summaryset}
\title{A function to create a gwasglue2 SummarySet object}
\usage{
create_summaryset(
  data = tibble(),
  metadata = NULL,
  tools,
  beta_col = "beta",
  se_col = "se",
  samplesize_col = "n",
  pvalue_col = "p",
  chr_col = "chr",
  position_col = "position",
  rsid_col = "rsid",
  effect_allele_col = "ea",
  other_allele_col = "nea",
  eaf_col = "eaf",
  id_col = "id",
  trait_col = "trait",
  ...
)
}
\arguments{
\item{data}{GWAS summary statistics. A tibble}

\item{metadata}{A list with metadata information. If NULL, it creates metadata with information retrieved from the dataset}

\item{tools}{Array of methods that gwasglue2 is going to convert the summarySet to (eg. "mr")}

\item{beta_col}{Name of column with effect sizes. The default is \code{"beta"}.}

\item{se_col}{Name of column with standard errors. The default is \code{"se"}.}

\item{samplesize_col}{Column name for sample size. The default is \code{"n"}.}

\item{pvalue_col}{Name of column with p-value. The default is \code{"p"}.}

\item{chr_col}{Column name for chromosome . The default is \code{"chr"}.}

\item{position_col}{Column name for the position. Together, with @param chr gives the physical coordinates of the variant. The default is \code{"position"}.}

\item{rsid_col}{Required name of column with variants rs IDs. The default is \code{"rsid"}.}

\item{effect_allele_col}{Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"ea"}.}

\item{other_allele_col}{Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \verb{"nea}.}

\item{eaf_col}{Name of column with effect allele frequency. The default is \code{"eaf"}.}

\item{id_col}{GWAS study ID column. The default is \code{"id"}.}

\item{trait_col}{Column name for the column with phenotype name corresponding the the variant. The default is \code{"trait"}}

\item{...}{Other columns}
}
\value{
A gwasglue2 SummarySet object
}
\description{
A function to create a gwasglue2 SummarySet object
}
\seealso{
\code{\link[=create_metadata]{create_metadata()}} to create a metadata object
}
