% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,DataSet-method}
\title{Get Methods for summaryset tibble (similar in Summaryset class)}
\usage{
getData(object, index)

\S4method{getData}{DataSet}(object, index)
}
\arguments{
\item{object}{A DataSet S4 object}

\item{index}{index of SummarySet within DataSet}
}
\value{
summaryset tibble
}
\description{
Get Methods for summaryset tibble (similar in Summaryset class)
}
