% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getMetadata}
\alias{getMetadata}
\alias{getMetadata,SummarySet-method}
\title{Get Method to retrieve the metadata stored in the SummarySet}
\usage{
getMetadata(summary_set)

\S4method{getMetadata}{SummarySet}(summary_set)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object.}
}
\value{
The gwasglue2 SummarySet metadata
}
\description{
Get Method to retrieve the metadata stored in the SummarySet
}
