% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{getSummarySet}
\alias{getSummarySet}
\alias{getSummarySet,DataSet-method}
\title{Get Method to retrieve the gwasglue2 SummarySet object}
\usage{
getSummarySet(object, index)

\S4method{getSummarySet}{DataSet}(object, index)
}
\arguments{
\item{object}{A gwasglue2 DataSet objec}

\item{index}{Index of gwasglue2 SummarySet objects within DataSet}
}
\value{
summarySet gwasglue2 SummarySet object
}
\description{
Get Method to retrieve the gwasglue2 SummarySet object
}
