% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{getVariants}
\alias{getVariants}
\alias{getVariants,SummarySet-method}
\title{Get Method to retrieve RSID/variants stored in the SummarySet}
\usage{
getVariants(summary_set)

\S4method{getVariants}{SummarySet}(summary_set)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object}
}
\value{
The RSID/variants
}
\description{
Get Method to retrieve RSID/variants stored in the SummarySet
}
\seealso{
Similar to \code{\link[=getRSID]{getRSID()}}
}
