% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise}
\alias{harmonise}
\title{Harmonise the alleles and effects between two summary sets
This is based o the harmonise function of TwoSampleMR
Harmonise function}
\usage{
harmonise(rsid, A1, A2, B1, B2, betaA, betaB, fA, fB, tolerance, action)
}
\arguments{
\item{rsid}{Array of rsids.}

\item{A1}{Array of effect alleles for trait A.}

\item{A2}{Array of other alleles for trait A.}

\item{B1}{Array of effect alleles for trait B.}

\item{B2}{Array of other alleles for trait B.}

\item{betaA}{Array of beta values for trait A.}

\item{betaB}{Array of beta values for trait B.}

\item{fA}{Array 0f frequency values for A1 allele.}

\item{fB}{Array of frequency values for B1 allele.}

\item{tolerance}{Tolerance value.}

\item{action}{Level of strictness in dealing with SNPs.
\itemize{
\item \code{action = 1}: Assume all alleles are coded on the forward strand, i.e. do not attempt to flip alleles
\item \code{action = 2}: Try to infer positive strand alleles, using allele frequencies for palindromes (default, conservative);
\item \code{action = 3}: Correct strand for non-palindromic SNPs, and drop all palindromic SNPs from the analysis (more conservative).
}}
}
\description{
Harmonise the alleles and effects between two summary sets
This is based o the harmonise function of TwoSampleMR
Harmonise function
}
