% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmoniseLD_method.R
\name{harmonise_ld}
\alias{harmonise_ld}
\title{Harmonise data against LD matrix}
\usage{
harmonise_ld(dataset, bfile = NULL, plink_bin = NULL)
}
\arguments{
\item{dataset}{The DataSet gwasglue2 object}

\item{bfile}{It corresponds to the path and prefix of the plink files used to build the LD correlation matrix.}

\item{plink_bin}{Path to the plink executable}
}
\value{
The DataSet gwasglue2 object harmonised
}
\description{
Function to create a LDmatrix gwasglue2 object and set the @slot ld_matrix u
}
