% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_ld.R
\name{harmonise_ld_dat}
\alias{harmonise_ld_dat}
\title{Harmonise LD matrix against summary data (now it just looks for overlaped variants. The harmonisation is done in ld_matrix_local)
harmonise_ld_dat() is based TwoSampleMR::harmonise_ld_dat()}
\usage{
harmonise_ld_dat(x, ld)
}
\arguments{
\item{x}{harmonised dataset}

\item{ld}{Output from ld_matrix}
}
\value{
List of dataset and harmonised LD matrix
}
\description{
Harmonise LD matrix against summary data (now it just looks for overlaped variants. The harmonisation is done in ld_matrix_local)
harmonise_ld_dat() is based TwoSampleMR::harmonise_ld_dat()
}
