% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merge Datasets}
\usage{
merge_datasets(...)
}
\arguments{
\item{...}{A list of DataSet S4 objects}
}
\value{
A DataSet S4 object  with input DataSets merged
}
\description{
Merge Datasets
}
