% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merge Datasets}
\usage{
merge_datasets(datasets)
}
\arguments{
\item{datasets}{A list of gwasglue2 DataSet objects}
}
\value{
A gwasglue2 DataSet object  with input DataSets merged
}
\description{
Merge Datasets
}
