% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{plot_gwasglue}
\alias{plot_gwasglue}
\title{Plot}
\usage{
plot_gwasglue(dataset, type, region)
}
\arguments{
\item{dataset}{gwasglue2 DataSet object}

\item{type}{Type of plot (Only available "manhattan" plots at the moment)}

\item{region}{Main title for the plot}
}
\value{
A plot
}
\description{
Plot
}
