% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setMetadata}
\alias{setMetadata}
\alias{setMetadata,SummarySet-method}
\title{Set Method to store metadata in the SummarySet}
\usage{
setMetadata(summary_set, metadata)

\S4method{setMetadata}{SummarySet}(summary_set, metadata)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object.}

\item{metadata}{A dataframe with metadata information}
}
\value{
gwasglue2 SummarySet object with metadata stored
}
\description{
Set Method to store metadata in the SummarySet
}
