% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{methods}
\name{setVariantid}
\alias{setVariantid}
\alias{setVariantid,SummarySet-method}
\title{Set Method to create an internal Variant ID for the SummarySet}
\usage{
setVariantid(summary_set)

\S4method{setVariantid}{SummarySet}(summary_set)
}
\arguments{
\item{summary_set}{A gwasglue2 SummarySet object}
}
\value{
A extra '"variantid"' column in the  GWAS summary statistics tibble. The \code{\link[=getSummaryData]{getSummaryData()}} can be used to retrieve it.
}
\description{
Set Method to create an internal Variant ID for the SummarySet
}
