% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{class}
\name{SummarySet-class}
\alias{SummarySet-class}
\title{An S4 class to represent the Summary Set}
\description{
An S4 class to represent the Summary Set
}
\section{Slots}{

\describe{
\item{\code{source}}{Eg.IEUopenGWAS (default NA).}

\item{\code{ss}}{A tibble with the GWAS summary statistics (default NA).}

\item{\code{metadata}}{A list with the metadata associated to ss (default NA).}

\item{\code{variants}}{The RSID/variants associated with ss (default NA).}

\item{\code{tools}}{The tools that gwasglue2 is going to convert to (default NA).}

\item{\code{mr_label}}{Exposure/Outcome (default NA).}
}}

