% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\name{SummarySet}
\alias{SummarySet}
\title{SummarySet function}
\usage{
SummarySet(sumstats)
}
\arguments{
\item{sumstats}{GWASb summary statistics}
}
\value{
A gwasglue2 SummarySet object.
}
\description{
SummarySet function
}
