% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\docType{class}
\name{SummarySet-class}
\alias{SummarySet-class}
\alias{SummarySet}
\title{An S4 class to represent the Summary Set}
\usage{
SummarySet(sumstats)
}
\arguments{
\item{sumstats}{GWAS summary statistics}
}
\value{
A gwasglue2 SummarySet object.
}
\description{
An S4 class to represent the Summary Set

SummarySet function
}
\section{Slots}{

\describe{
\item{\code{ss}}{A tibble with the GWAS summary statistics (default NA).}

\item{\code{metadata}}{A list with the metadata associated to ss (default NA).}

\item{\code{variants}}{The RSID/variants associated with ss (default NA).}

\item{\code{attributes}}{Attributes of the SummarySet. Eg. MR label Exposure/Outcome (default NA).}
}}

\keyword{internal}
