% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryset.R
\name{SummarySet}
\alias{SummarySet}
\title{SummarySet function}
\usage{
SummarySet(ss, traits, variants, tools)
}
\arguments{
\item{ss}{It uses createSumset function to call ieugwasR and fill @slot ss}

\item{traits}{ID of GWAS studies to query}

\item{variants}{Array of SNPs rsids}

\item{tools}{Array of methods that gwasglue2 ids going to convert the summarySet to (eg. "mr")}
}
\value{
A SummarySet S4 object.
}
\description{
SummarySet function
}
